session처리 완료, doAdd에 작성자 정보 저장, modify, delete 권한체크
ResultData 변수 추가해서 메서드들 수정해야함

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -2,6 +2,8 @@ package com.example.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -26,7 +28,11 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData<Article> doAdd(String title, String body) {
+	public ResultData<Article> doAdd(HttpSession session, String title, String body) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
 		
 		if (Util.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요.");
@@ -36,7 +42,9 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "내용을 입력해주세요.");
 		}
 		
-		articleService.writeArticle(title, body);
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		
+		articleService.writeArticle(loginedMemberId, title, body);
 		
 		int id = articleService.getLastInsertId();
 		
@@ -50,7 +58,7 @@ public class UsrArticleController {
 		List<Article> articles = articleService.getArticles();
 		
 		if (articles.size() == 0) {
-			return ResultData.from("F-1", "게시물이 존재하지 않습니다.");
+			return ResultData.from("F-1", "게시글이 존재하지 않습니다.");
 		}
 		
 		return ResultData.from("S-1", "게시글 목록", articleService.getArticles());
@@ -63,7 +71,7 @@ public class UsrArticleController {
 		Article foundArticle = articleService.getArticleById(id);
 		
 		if (foundArticle == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다.", id));
 		}
 		
 		return ResultData.from("S-1", "%d번 게시글 입니다.", foundArticle);
@@ -71,30 +79,46 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession session, int id, String title, String body) {
 		
-		Article foundArticle = articleService.getArticleById(id);
-		
-		if (foundArticle == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
 		}
 		
-		articleService.modifyArticle(id, title, body);
-		return ResultData.from("S-1", Util.f("%d번 게시물을 수정했습니다.", id));
+		if (Util.empty(id)) {
+			return ResultData.from("F-1", "글 번호를 입력해주세요.");
+		}
+		
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		
+		return articleService.modifyArticle(loginedMemberId, id, title, body);
 	}
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
+		
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+		
+		if (Util.empty(id)) {
+			return ResultData.from("F-1", "글 번호를 입력해주세요.");
+		}
 		
 		Article foundArticle = articleService.getArticleById(id);
-	
+		
 		if (foundArticle == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+			return ResultData.from("F-2", Util.f("%d번 게시글은 존재하지 않습니다.", id));
+		}
+		
+		if ((int) session.getAttribute("loginedMemberId") != foundArticle.getMemberId()) {
+			return ResultData.from("F-B", "해당 게시글에 대한 권한이 없습니다.");
 		}
 		
 		articleService.deleteArticle(id);
-		return ResultData.from("S-1", Util.f("%d번 게시물을 삭제했습니다..", id));
+		return ResultData.from("S-1", Util.f("%d번 게시글을 삭제했습니다.", id));
 	}
 	
 }
